@@ -1,42 +1,23 @@
-import { useState } from "react";
+import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-import decoration from '../../resources/img/vision.png';
-import AppBanner from "../appBanner/AppBanner";
-import ComicsList from "../comicsList/ComicsList";
+import { MainPage, ComicsPage, Page404, SingleComicPage } from '../pages';
+
 
 const App = () => {
-	const [selectedChar, setSelectedChar] = useState(null);
-
-
-	const onCharSelected = (id) => {
-		setSelectedChar(id);
-	}
-
 	return (
+<Router>
 		<div className="app">
 			<AppHeader />
 			<main>
-				{/* <ErrorBoundary>
-					<RandomChar />
-				</ErrorBoundary>
-				<div className="char__content">
-					<ErrorBoundary>
-						<CharList onCharSelected={onCharSelected} />
-					</ErrorBoundary>
-					<ErrorBoundary>
-						<CharInfo charId={selectedChar} />
-					</ErrorBoundary>
-				</div> */}
-				<AppBanner />
-				<ComicsList />
-
-				<img className="bg-decoration" src={decoration} alt="vision" />
+			<Routes>
+				<Route path="/" exact element={<MainPage/>}/>
+				<Route path="/comics" element={<ComicsPage/>}/>
+				<Route path="/comics/:comicId" element={<SingleComicPage/>}/>
+				<Route path='*' element={<Page404/>}/>
+			</Routes>		
 			</main>
 		</div>
+</Router>
 	)
 }
 
