@@ -1,4 +1,5 @@
 import { useEffect, useState } from 'react';
+import { Link } from 'react-router-dom';
 import './comicsList.scss';
 import useMarvelService from '../../services/MarvelService';
 import ErrorMessage from '../errorMessage/ErrorMessage';
@@ -43,11 +44,11 @@ const ComicsList = () => {
 			return (
 				<li className="comics__item"
 					key={i}>
-					<a href="#">
+					<Link to={`/comics/${item.id}`}>
 						<img src={item.thumbnail} alt={item.title} className="comics__item-img" />
 						<div className="comics__item-name">{item.title}</div>
 						<div className="comics__item-price">{item.price}</div>
-					</a>
+					</Link>
 				</li>
 			)
 		})
